Redesigned TetrisBoard and implemented block downward movement.
